+++ AkCodeWorkSpaceCreator.sln
@@ -30,6 +30,7 @@ Global
 		{68D14A80-2900-4635-BA42-2F5BB9282F5D}.Release|x86.ActiveCfg = Release|Any CPU
 		{68D14A80-2900-4635-BA42-2F5BB9282F5D}.Release|x86.Build.0 = Release|Any CPU
 		{02E507B6-88BA-45B3-A9A9-9073A995BF04}.Debug|Any CPU.ActiveCfg = Debug
+		{02E507B6-88BA-45B3-A9A9-9073A995BF04}.Debug|Any CPU.Build.0 = Debug
 		{02E507B6-88BA-45B3-A9A9-9073A995BF04}.Debug|x64.ActiveCfg = Debug
 		{02E507B6-88BA-45B3-A9A9-9073A995BF04}.Debug|x86.ActiveCfg = Debug
 		{02E507B6-88BA-45B3-A9A9-9073A995BF04}.Release|Any CPU.ActiveCfg = Release
